feat(runtime-c-api) Extract the `error` module.
